Switch to new database
USDA slightly changed the format of the SR csv files in the latest
version. This is the last version of the SR database that will be
published, but I thought it was worth upgrading since this new format is
similar to the new USDA database formats and to correct any typo (at
least one known typo in the last version). I also updated the column
names.